Move to a rolling memoization process, remove freeze
There is no freeze mechanism anymore. Each time the circuit is altered,
a "timestamp" counter is increased, and the last alteration time is then
set to that.
Each memoization is associated with a timestamp. When a memoized result
is accessed, it is discarded and recomputed if it has expired (ie. its
timestamp is older than the current alteration time).

WARNING: alteration timestamps are not consistent between circuits, each
circuit has its timestamp counter. This can easily be circumvented by
making the counter "static", but I see no point in that.

WARNING: the timestamp counter is a 64 bits unsigned integer (defined
as size_t). If you do *really* heavy computation on a single circuit
(say, a root group), it might overflow at some point and cause weird
bugs.